requirements.txt
pandas==2.1.0
scikit-learn==1.3.0
numpy==1.24.3
Flask==3.0.2
joblib==1.3.2
gunicorn==21.2.0
python-dotenv==1.0.0 